refine symbols tables and reformat
