+++ server/parser/asylexer.py
@@ -1,5 +1,6 @@
 from .ply.lex import lex
 
+
 def _find_column(input, token):
     line_start = input.rfind("\n", 0, token.lexpos) + 1
     return (token.lexpos - line_start) + 1
